+++ src/main/kotlin/ar/edu/unq/agiletutor/webservice/TutorRestService.kt
@@ -65,7 +65,7 @@ class TutorRestService {
 
     /**get tutor by id**/
     @GetMapping("/api/tutor/{id}")
-    fun tutorById(@PathVariable("id") id: Int): ResponseEntity<*> {
+    fun tutorById(@PathVariable("id") id: Long): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             val tutorView = TutorDTO.desdeModelo(tutorService.findByID(id))
@@ -101,7 +101,7 @@ class TutorRestService {
 
     /** Update*/
     @PutMapping("/api/tutor/{id}")
-    fun update(@PathVariable("id") id: Int, @RequestBody entity: TutorRegisterDTO): ResponseEntity<*> {
+    fun update(@PathVariable("id") id: Long, @RequestBody entity: TutorRegisterDTO): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             val userview = tutorService.update(id, entity)
@@ -121,13 +121,13 @@ class TutorRestService {
 
     /**Delete tutor by id*/
     @DeleteMapping("/api/tutor/{id}")
-    fun deleteTutorById(@PathVariable("id") id: Int): ResponseEntity<*> {
+    fun deleteTutorById(@PathVariable("id") id: Long): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             tutorService.deleteById(id)
-            val resultado: MutableMap<String, Int> = HashMap()
+            val resultado: MutableMap<String, Long> = HashMap()
             resultado["succesfully tutor deleted with iD"] = id
-            response = ResponseEntity.ok().body<Map<String, Int>>(resultado)
+            response = ResponseEntity.ok().body<Map<String, Long>>(resultado)
 
         } catch (e: Exception) {
             ResponseEntity.status(404)
@@ -141,7 +141,7 @@ class TutorRestService {
 
     /**Courses From a Tutor*/
     @GetMapping("/api/tutor/courses/{id}")
-    fun coursesFromATutor(@PathVariable("id") id: Int): ResponseEntity<*> {
+    fun coursesFromATutor(@PathVariable("id") id: Long): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             val courses = tutorService.coursesFromATutor(id).map { CourseDTO.desdeModelo(it) }
@@ -163,7 +163,7 @@ class TutorRestService {
 
     /**Students From a Tutor*/
     @GetMapping("/api/tutor/students/{id}")
-    fun studentsFromATutor(@PathVariable("id") id: Int): ResponseEntity<*> {
+    fun studentsFromATutor(@PathVariable("id") id: Long): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             val students = tutorService.studentsFromATutor(id).map { StudentDTO.desdeModelo(it) }
@@ -187,7 +187,7 @@ class TutorRestService {
     /** Move a student to Another Course*/
     @PutMapping("/api/tutor/students/move/{id}/{id_course}")
     fun moveAStudentIntoAnotherCourse(
-            @PathVariable("id") id: Int, @PathVariable("id_course") id_course: Int
+            @PathVariable("id") id: Int, @PathVariable("id_course") id_course: Long
     ): ResponseEntity<*> {
         var response: ResponseEntity<*>?
 
@@ -209,7 +209,7 @@ class TutorRestService {
 
     /** Absent Message From a Tutor*/
     @GetMapping("/api/tutor/absentmessage/{tutorId}")
-    fun absentMessageFromTutor(@PathVariable("tutorId") tutorId: Int): ResponseEntity<*> {
+    fun absentMessageFromTutor(@PathVariable("tutorId") tutorId: Long): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             val notifyer = tutorService.findByID(tutorId).notifyer
@@ -230,7 +230,7 @@ class TutorRestService {
 
     /** Modify Absent Message From a Tutor*/
     @PutMapping("/api/tutor/absentmessage/{tutorId}")
-    fun updateAbsentMessageFromTutor(@PathVariable("tutorId") tutorId: Int, @RequestBody absentMessageDataDTO: AbsentMessageDataDTO): ResponseEntity<*> {
+    fun updateAbsentMessageFromTutor(@PathVariable("tutorId") tutorId: Long, @RequestBody absentMessageDataDTO: AbsentMessageDataDTO): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             val notifyer = tutorService.findByID(tutorId).notifyer
@@ -251,7 +251,7 @@ class TutorRestService {
 
     /**Students To Notify From a Tutor*/
     @GetMapping("/api/tutor/toNotify/{tutorId}")
-    fun studentsToNotify(@PathVariable("tutorId") tutorId: Int): ResponseEntity<*> {
+    fun studentsToNotify(@PathVariable("tutorId") tutorId: Long): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             val students = tutorService.studentsToNotifyFromTutor(tutorId)
@@ -269,7 +269,7 @@ class TutorRestService {
 
     /**remove an absent to notify*/
     @PostMapping("/api/tutor/absent/{tutorId}/{studentId}")
-    fun login(@PathVariable("tutorId") tutorId: Int, @PathVariable("studentId") studentId: Int): ResponseEntity<*> {
+    fun login(@PathVariable("tutorId") tutorId: Long, @PathVariable("studentId") studentId: Long): ResponseEntity<*> {
         var response: ResponseEntity<*>?
         try {
             tutorService.removeAnStudentFromNotification(tutorId, studentId)
